% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{check_link}
\alias{check_link}
\title{check if link work}
\usage{
check_link(link_in)
}
\arguments{
\item{link_in}{link to be tested (character)}
}
\value{
dataframe with result
}
\description{
check if link work
}
\examples{
check_link("www.google.com")
}
