% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todos.R
\name{find_todos}
\alias{find_todos}
\title{Find "TODO:"  in files}
\usage{
find_todos(f_in)
}
\arguments{
\item{f_in}{file to be tested}
}
\value{
a flag (TRUE/FALSE)
}
\description{
Find "TODO:"  in files
}
\examples{
  example_file <- fs::path(
system.file('extdata/examples/example_RMD.Rmd', package = 'textcheck')
)
flag <- find_todos(example_file)
}
